fix(Switch): Move NewSwitch to Switch
